backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
python-dotenv==1.0.1
duckdb==1.0.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pyodbc==5.1.0
